Add ORSR records checker
For a given identifier (IČO) it retrieves an array of results for all records found for the identifier on orsr.sk

It skips records which were moved to another court/district.

Each result is a hash with ok/missing keys, where the key holds an array of names of stakeholders.