Do not assume message of the deadline component
This commit changes the way how deadline is communicated. Instead of a pre-defined message, you have a possibility to use :remaining_days: as a placeholder within your text. When rendering it, it will be replaced by "1 deň", "M dni" or "N dní" depending on the actual number of days to reach the deadline.

An example of usage:

```HTML
<deadline time='12-27'>
 <before>Založte si živnosť najneskor za :remaining_days: podľa <a href=/nieco> tohto návodu</a>.</before>
 <after>Toto ste už nestihli, ale aj tak si založte živnosť podľa <a href=/nieco> tohto návodu</a>.
</deadline>
```

refs. #147