Do not pre-define message of the deadline component (#543)
* Do not assume message of the deadline component

This commit changes the way how deadline is communicated. Instead of a pre-defined message, you have a possibility to use :remaining_days: as a placeholder within your text. When rendering it, it will be replaced by "1 deň", "M dni" or "N dní" depending on the actual number of days to reach the deadline.

An example of usage:

```HTML
<deadline time='12-27'>
 <before>Založte si živnosť najneskor za :remaining_days: podľa <a href=/nieco> tohto návodu</a>.</before>
 <after>Toto ste už nestihli, ale aj tak si založte živnosť podľa <a href=/nieco> tohto návodu</a>.
</deadline>
```

refs. #147

* Use handlebars style for remaining days templating

* Handle recurring events with deadline long time ago as upcoming

We treat recurring past due events which had deadline more than a half a year ago from today as upcoming, as the next occurrence is actually closer to today than the previous one.

Co-authored-by: Jano Suchal <jan.suchal@gmail.com>