Validate provided URL when linking journey to law
SlovLexLink is able to check basic validity of the provided URL via regexp and throws an error if pattern does not match. If it does match, SlovLexHelper attempts to download & parse the law document to validate that the link points to a real law and there was no typo in law identification numbers.