Show journey as outdated if it was updated after check date
If there is a change in law version text, checksum will change and `updated_at` column of this law version will reflect the date when this change was detected. If a linked journey was checked before the date when law texts changed, we consider this journey as outdated (needs to be checked again).