Enhance submissions cleanup process
Submissions are being pruned from the database according to their expires_at timestamp. Submissions which are not delivered by email are deleted after 20 minutes. Submissions which are to be sent via email do not have any initial expiration timestamp set so that any accidental outage/delay of email delivery worker does not lead to data loss.

Once the email is sent, the expiration time is set to 20 minutes to allow further interactions with the submission (#finish) within that period.