Be able to know which Journey is outdated (#534)
* Add SlovLex link class

* Add basic law model / migration

* Add editing of legal definitions of journeys in admin

* Move law id & creation into model

So that it covers update method as well.

* Rename to Law link in UI

* Point law link detail to actual external link

* Add LawVersion model

* Add sync version for jobs

* Retrieve & show journey law check status

* Call job with perform_later

* Validate provided URL when linking journey to law

SlovLexLink is able to check basic validity of the provided URL via regexp and throws an error if pattern does not match. If it does match, SlovLexHelper attempts to download & parse the law document to validate that the link points to a real law and there was no typo in law identification numbers.

* Show warning message on journeys with outdated check

* Show journey as outdated if it was updated after check date

If there is a change in law version text, checksum will change and `updated_at` column of this law version will reflect the date when this change was detected. If a linked journey was checked before the date when law texts changed, we consider this journey as outdated (needs to be checked again).

* Refactor + show status on law links of a journey screen

* Schedule weekly law check job

* Fix rake task

* Update clock.rb

* Fix handling of nil values

* Handle few more nils of blank state

* Fix typo

* Do not show problem with laws if there are none for a journey

---------

Co-authored-by: Michal Barla <michal.barla@gmail.com>
Co-authored-by: Jano Suchal <jan.suchal@slovensko.digital>